Correct import module errors and eslint errors
