+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Awesome Books</title>
+    <title>Awesome Books with ES 6</title>
     <link rel="stylesheet" href="./index.css" />
   </head>
   <body>
