+++ modules/date-time.js
@@ -2,7 +2,7 @@ import { DateTime } from './luxon.js';
 
 const myDateTime = () => {
   const myDocument = document.getElementById('date-time');
- 
+
   const dateTimeUpdate = () => {
     const dateTimeCurrent = DateTime.now().toLocaleString(
       DateTime.DATETIME_FULL_WITH_SECONDS,
