+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "awesome_books",
+  "name": "Awesome-Books-with-ES6",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -8,28 +8,16 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/mulugithub/Awesome_Books.git"
+    "url": "git+https://github.com/anyars-encarta/AwesomeBooks-with-ES6.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/mulugithub/Awesome_Books/issues"
-  },
-  "homepage": "https://github.com/mulugithub/Awesome_Books#readme",
-  "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.27.5",
-    "hint": "^7.1.10",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^21.0.0",
-    "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.21.0"
-  },
+  "homepage": "https://github.com/anyars-encarta/AwesomeBooks-with-ES6/#readme",
   "dependencies": {
     "full-icu": "^1.5.0",
-    "luxon": "^3.3.0"
+    "luxon": "^3.3.0",
+    "parcel-bundler": "^1.12.5",
+    "three": "^0.154.0"
   }
 }
